[Develop] Add data integrity checks before removing backup directories (#2727)
* feat: Add data integrity checks before removing backup directories
  * Updated `restore_home_shared_data.rb` to include a diff check before removing /tmp/home/ directory.
  * Updated `restore_internal_use_shared_data.rb` to include a diff check before removing /tmp directories for each internal shared directory.
  * Updated `config_default_user_home.rb` to include data integrity checks during the process of moving the cluster user's default home directory.
  * Added new Spec tests.

* This update prevents potential data loss or inconsistency during the restoration process.

+++ cookbooks/aws-parallelcluster-environment/recipes/init/restore_home_shared_data.rb
@@ -19,13 +19,22 @@ if node['cluster']['node_type'] == 'HeadNode'
   # This is necessary to preserve any data in these directories that was
   # generated during the build of ParallelCluster AMIs after converting to
   # shared storage and backed up to a temporary location previously
-  # Remove the backup after the copy is done
+  # Before removing the backup, ensure the data in the new home is the same
+  # as the original to avoid any data loss or inconsistency. This is done
+  # by using rsync to copy the data and diff to check for differences.
+  # Remove the backup after the copy is done and the data integrity is verified.
   bash "Restore /home" do
     user 'root'
     group 'root'
     code <<-EOH
       rsync -a --ignore-existing /tmp/home/ /home
-      rm -rf /tmp/home/
+      diff_output=$(diff -r /tmp/home/ /home)
+      if [ $? -eq 0 ]; then
+        rm -rf /tmp/home/
+      else
+        echo "Data integrity check failed comparing /home and /tmp/home: $diff_output"
+        exit 1
+      fi
     EOH
   end
 end
